+Added GLOBAL_INDEXING run-time parameter
  Added a new run-time parameter, GLOBAL_INDEXING, to use use global array
indexing in MOM6.  Something similar already works in SIS2. By default this is
false and answers are unchanged.  Answers should not be altered by this setting,
but because it has not been exercised, at least some of the MOM6 test cases fail
spectacularly when this is enabled.  Also, the location of the call to
hor_index_init had to be relocated after several get_param calls, which alters
the order of some lines when the parameters are logged.  By default all test
cases are bitwise identical, but new MOM_parameter_doc files have been added.