+Moved WRITE_GEOM into initialize_MOM
  Moved the run-time parameter controlling whether the horizontal geometry and
vertical grid files are written up into initialize_MOM.  This both places this
parameter next to similar parameters (like SAVE_INITIAL_CONDS), and facilitates
the separation of the initialization of the vertical grid from the horizontal
grid.  Because it changes the order in which variables are logged in the
MOM_parameter_doc files, new versions needed to be added for the test cases.
All answers are bitwise identical.