Initialize G%bathyT to 0.
  Changed the initialization value of G%bathyT from Angstrom_Z to 0.  This value
is updated via halo updates, but can persist on the edges of the global domain.
This could change where masks are applied around the edges, depending on the
values of MINIMUM_DEPTH and MASKING_DEPTH (with the new value more likely to
give the desired answer), but it happens not to change the answers in any of the
MOM6 test cases.