+* Revised create_depth_list
  Rewrote the section of create_depth_list that sets up a list of all depths.
This list will now be guaranteed to have the same order of cells as if had been
created by a single-PE run.  This can change the reported energies if
CALCULATE_APE is true and the topography and area are internally generated or
read from a 64-bit file, but it just so happens not to change any of the
existing reference solutions.  In addition, dOxyGen comments were added to
create_depth_list.
This commit addresses and closes Issue NOAA-GFDL/MOM6-Examples#82