Changes in the ocean initialization
We now initialize the ocean layers consistent with the averaged density between layers. Ice shelf thickness (soon to be mass) is read during the initialization and the interfaces are displaced accordingly. The problem is that the pressure at the surface (p_surf) is not consistent with that, so the model crashes. This needs to be fixed!