+Added finish_MOM_initialization.
  Added a new subroutine, finish_MOM_initialization, that is called at the
beginning of the first timestep, after the first forcing field is called.
This allows the model to be updated to reflect fields like surface_pressure
that were not known yet during initialize_MOM.  For now, there is nothing
actually in this routine, but that will change.  All answers are bitwise
identical, although a new public interface was added.