*Rewrote find_root for set_grid_metrics_mercator
  Rewrote the find_root routine that is used by set_grid_metrics_mercator to
work robustly and efficiently even for poor initial guesses, allowing the
latitude of all points to be calculated using the same (arbitrary) initial guess
in the middle of the domain, and removing a dependence on the order in which
they are calculated.  This changes the answers at the level of round-off for any
cases that use GRID_CONFIG = "mercator", including benchmark and MESO_025_63L.
New reference solutions have been checked in for the benchmark test case (the
MESO case still works, but we are no longer providing reference solutions).