Corrected agressive nonblocking sends in step_MOM
  Reordered the complete_group_pass for the surface pressure and wind stresses
relative to their use or to the chksum calls to evaluate their validity in order
to avoid falsely inconsistent halo values or (in the case where INTERP_P_SURF =
True and NONBLOCKING_UPDATES=True) the possibility of actual errors.  Also added
a call to check_redundant for the wind stresses when DEBUG=True.  Answers are
the same in all test existing test cases.