*Call fill_symmetric_edges for taux and tauy.
  Added calls to fill_symmetric_edges for taux and tauy so that they are given
proper data in the symmetric edge regions.  In the case of coupled or ice-ocean
model simulations with MOM6 using symmetric memory, this corrects an obvious
parallelization bug.  In the case of ocean-only simulations, this cleans up code
that had been filling in these values properly, but in a rather cumbersome way
via an extra set of halo updates.  This does change answers in coupled runs
using symmetric memory, but not in any existing MOM6 test cases.
This commit addresses and closes Issue NOAA-GFDL/MOM6-Examples#282.