*Corrected ocean-only fill_symmetric_edges call
  Corrected second argument in fill_symmetric_edges call in
wind_forcing_from_file.  This is an obvious bug, introduced in
NOAA-GFDL/MOM6@17ec5f5ef8648a3bcc974f5dba1ace38519f2454, but it was missed
in testing because it only impacts symmetric memory ocean-only test cases.
Simulations with this configuration would fail with an error message very early
in the run.