+Added initialization option to fit salinity
  Added the option to initialize_temp_salt_fit to use prescribed temperatures
and fit the salinity to the prescribed layer densities, as well as the original
code that did the opposite.  This mode is selected with the new run-time
parameter FIT_SALINITY. Also made changes to allow trim_for_ice to work without
using an ALE mode.  All answers are bitwise identical, but there are changes to
the MOM_parameter_doc.all files for cases with TS_CONFIG=fit.