Avoid division by zero in continuity_PPM
  Added code to avoid division by a 0 value of visc_rem_lim in the unlikely case
where the input velocity is finite but the viscosity completely eliminates the
flow everywhere in the column.  As this can only happen for a masked or
completely massless column, this should not happen, but it was triggered by a
bug in setting the initial velocities in the mixedlayer_restrat_2d test case
with symmetric memory.  This does not change answers in any case that worked
before.  This update partially address MOM6 Issue #299.