Added a passive tracer to the ice shelf melt water
I am now adding a second passive tracer (m=2) based on the ice shelf melt water field. To do so I first included the melt field in the structure that contains pointers to the boundary forcing (defined in MOM_forcing_type.F90) so this field can be accessed from ISOMIP_tracer.F90. This field is passed to forcing structure in MOM_ice_shelf.F90. Finally, the tracer is set in ISOMIP_tracer_column_physics (ISOMIP_tracer.F90).

PS: this needs to be improved. It is still not working the way it should.