Completed and tested MOM_transcribe_grid
  Completed the two subroutines copy_MOM_grid_to_dyngrid and
copy_dyngrid_to_MOM_grid in MOM_transcribe_grid, and verified that they work as
intended. Made some minor changes (to or from pointers) to some of the types
inside of the dyn_horgrid_type and to arguments in MOM_dyn_horgrid.F90. Added
code in MOM_initialization to test the ability to copy back and forth between a
ocean_grid_type and a dyn_horgrid_type via the routines in MOM_transcribe_grid.
(These tests are hard coded to be off; the answers do not change, but some
additional logging to MOM_parameter_doc.all occurs with these tests.) All
answers are bitwise identical, but there have been minor changes to some
previously unused interfaces.