+Created diag_mediator_infrastructure_init
  Split out the FMS infrastruture initialization that had occurred within
diag_mediator_init into a new subroutine, diag_mediator_infrastructure_init.
This will allow diag_mediator_init to be called whenever it makes sense for MOM6
without causing problems for the FMS system as a whole.  The reason for this
change is that the modifications introduces recently with commit
NOAA-GFDL/MOM6@a836bcb90c473c26022885232ca63e4350b8a723 were causing problems
with the new Verona release of FMS with the gnu compiler (although there was no
issue with the Ulm version).  There is a new subroutine that must be called, and
a now unused optional argument is removed form diag_mediator_init.  All answers
appear to be bitwise identical, although testing has not been done with all
compilers and test cases.