+Removed ocean_grid_type form ALE_init.
  The ALE initialization routines were making minimal use of data from the
ocean_grid_type, so instead this data is taken from the vertical grid type, and
the maximum depth is explicitly passed in as an argument to both ALE_init and
ALE_initRegridding.  In addition, the horizontal grid type is removed as an
(unused) argument to ALE_writeCoordinateFile.  All solutions are bitwise
identical, although several interfaces have changed.