Doxygen macro expansion and variable grouping
We expand only the macro ALLOCABLE_ to the Fortran keyword
'allocatable'. This cleans up some type definitions, where doxygen
thought that allocable_ was a variable name and got confused about what
the type members actually were.

We also turn on variable grouping with DISTRIBUTE_GROUP_DOC
If there's a set of variables that all do the same thing, such as
diagnostic identifiers, they can be grouped and the same docstring will
be used for all members. For example:

    !>@{
    !! Diagnostic identifiers
    integer :: id_u = -1, id_v = -1, id_h = -1
    !>@}