Added sponge layers when using layer mode
Now the ISOMIP initialization allows the implementation of sponge layer when using layer mode. This is done by reading a NETCDF file, for example the IC file saved by the model, that specifies eta, T and S. The T and S values most be consistent with the target density of the layers.