+Use dyn_horgrid_type in fixed initialization
  Made extensive changes to replace the ocean_grid_type with a dyn_horgrid_type
during the fixed initialization of MOM6.  This is a big step toward being able
to share MOM6 fixed initialization software (of masks, topography, and the
Coriolis parameter) with other components that have their own layouts.  This
changes the type of a number of arguments, and adds some arguments in a few
cases of ..._initialize_topography files (for standardization).  Also added
dOxyGen comments for some subroutines.  All answers are bitwise identical.