Deallocate G%Domain in destroy_dyn_horgrid
  Added calls to deallocate G%Domain and G%Domain%mpp_domain in
destroy_dyn_horgrid.  This eliminates a potential memory leak. It had not been
done before because there was some concern that these pointers might have been
set to point to an element of another type, but this is now all being done via a
copy, so they can (and should) be deallocated.  All answers are bitwise
identical.