Larger allocations for diffusivity diagnostics
Some of the diagnostics in MOM_set_diffusivity (Nb, TKE_itidal) have
arrays that are allocated in the main control structure. These were
being allocated on (isc:iec,jsc:jec), which was causing an out-of-bounds
array access in FMS upon diagnostic post (due to the masks being
larger).

We allocate these arrays on the slightly larger domain (isd:ied,jsd:jed)
to fix this, pending a change to post_data which checks that the posted
array is compatible with its mask.