Added parameter MELTING_CUTOFF_DEPTH
This sets the depth above which the melt is set to zero (required for the ISOMIP test case). I've used pressure (mass of ice shelf * RHO_0 * g) to find out whether lprec needs to be set to zero. Since RHO_0 is used, this is an approximation which works fine in this case.