Added a new paramter in the ice shelf code and modified isomip_tracer
Added a new parameter (MELTING_CUTOFF_DEPTH) in MOM_ice_shelf.F90 that controls the above which the melt is set to zero.

Modified ISOMIP_tracer.F90 so that only one tracer is now injected under the ice shelf. There was another tracer being injected in the sponge layer (code for that is not commented), but that did not wrok in layer mode (it needs to be modified to allow sponge in addition to sponge_ale).