Compute velocities (and add to state) in the ML when *not* using BULKMIXLAYER
When BULKMIXLAYER is not defined, both u and v are averaged over the uppermost depth_ml fluid. Before, just tracers were being averaged over depth_ml. This change is needed to correctly compute u and v in the ML when layer thicknesses are vanished (i.e., under an ice shelf).