Average u and v in the uppermost depth_ml fluid
When BULKMIXLAYER is *not* defined, both u and v are now averaged over the uppermost depth_ml fluid. This is done in the subroutine "calculate_surface_state". Before, just tracers were being averaged over this depth. This change is needed to correctly compute u and v in the mixed layer (ML) when layer thicknesses are vanished (i.e., under an ice shelf) and *ALE* mode is used. The averaged ML fields are passed to the surface state structure, which is then used to compute melting in the ice shelf module.