Implemented "interp_method = 'none'" for vector/q-point locations
To avoid fregrid interpolation components of vectors or fields that
cannot be interpolated using conservative methods a netcdf attribute
"interp_method" has been asked for. This is a non-CF attribute that
tells fregrid how to interpolate the field. The specification suggets
specifying the order of interpolation which is not a forward compatible
approach. In order to facilitate post-processing for CMIP6 we are
partially implementing this method by only attaching interp_method='none'
to for vector-components and non-interpolatable fields.

- Type axes_grp now contains a flag "is_h_point" to efficiently discern
  staggering location on the C-grid.
- Corrected comments for optional argument "interp_method" to
  register_diag_field which used to be "no clue".
- Updated registrations in MOM.F90:write_static_fields() to specify
  interp_method explicitly for static diagnostics.
- No answer changes.