Fields ustar_berg, area_berg and mass_berg have been passed from the iceberg model to SIS2, through the coupler to the ocean model. The ocean model has been edited to allow these fields to be passed to it when the flag ICEBERGS_APPLY_RIGID_BOUNDARY=.True.
The iceberg surface fields are added to the ice shelf fields (if the ice shelf is running). The combined ice shelf /iceberg fields can now affect the upper ocean boundary conditions. This is done by adding area_berg to frac_shelf_h, and adding ustar_berg to ustar_shelf. The mass is also used to calculate the shelf rigidity