Updates for KH in hor_visc and CDRAG in MEKE
Update to allow read of KH from external file in MOM_hor_visc
Update to allow differentiation of CDRAG and CDRAG_MEKE in MOM_MEKE (not sure of utility of this)
MOM_forcing_type : added checks for presence of optional argument before calling subroutine with it.
(I believe it is handled ok in the called subroutine but just in case. You can take it or leave it.)
MOM.F90 (I don't remember why this was needed but it helped with a crash. Maybe to do with AMIP? Ask Zhi if you need to.)