Added option to calculate sea level using local density.
An option has been added to convert the surface pressure to sea level, using the locally calculated density. By default, the code uses the boussinesq density Rho0. When the flag CALC_RHO_FOR_SEA_LEVEL is True, the code calculates a local density to use for the conversion. The local density is calculated using the SST, SSS at p_surf/2. This is still an approximation, but does better than using the boussinesq density.

This flag lowers the shadow of iceberg which can be seen in the sea level fields.