Berg_area_threshold added to determine if sub-icebergs fluxes must be set to zero
A parameter called Berg_area_threshold has been added. When the parameter is zero, the fluxes beneath the ice shelf are set to zero. When the parameter is non-zero, the fluxes are set to zero if the fraction of the grid cell area covered by icebergs is greate than berg_area_threshold