Added parentheses in advect_x
  Added parentheses to the calculation of which branch to use in calculating
the zonal mass flux of a particular iteration in advect_x.  This choice of
parentheses matches what is currently done in advect_y and is the pairing that
is preferred for this calculation.  This could have changed answers at the level
of roundoff, but all answers in the test cases are bitwise identical.