+Removed optional arguments from chksum
  Removed the index range optional arguments from the very old style calls to
chksum that resolved to chksum_2d and chksum_3d.  These arguments are dangerous
because the index ranges are changed upon entry to these routines and they
were not used in any MOM6 code.  Although the interfaces are changed, no other
code needed to be changed and all solutions are bitwise identical.