Simplify calculation of do_j_tr to use stencil in advect_y
We can collapse two indentical loops with slightly different bounds,
guarded by conditions to a single loop. This depends only on the
stencil for which slope calculations are required (i.e. when using
PLM, only the slopes immediately adjacent the face exhibiting a flux
are required, whereas with PPM an additional point is required on each
side, as signified by the stencil of the scheme).