Enable control of diffusivity parameters for ADAPT
We add three parameters: ADAPT_TIME_RATIO, ADAPT_ZOOM_DEPTH and
ADAPT_ZOOM_COEFF. The time ratio is the ratio of an ALE timestep to a
nominial grid timescale. The larger this value, the stronger the grid
diffusivity. The zoom depth is the constant parameter determining the
strength of near-surface zooming in the grid diffusivity, which has a
relative strength to background diffusivity controlled by the zoom
coefficient.