Use \namespace instead of \class for documentation
We document Fortran modules as Doxygen "namespaces", which seem to be
a pretty good analogue. However, the files in src/user/ were
documenting modules as "classes", which was causing fragmentation and
confusion in the Doxygen output.

Additionally, because of the case-insensitive nature of Fortran, we
need to use the lowercase module name for a documentation
block (`\namespace module_name`), otherwise Doxygen thinks you are
documenting a separate module and generates an additional output file
accordingly.