Continue work on making sure that remapped diags are done correctly
In thickness_diffuse, some lines are rearranged to make sure that the diags
are posted on the grid before layer thicknesses are changed. The call to
update the remapping grids is done after those diags.

In diag_update_remap_grids, an optional argument can now be passed in if
the appropriate layer thicknesses are not the current h. This is primarily
relevant for the umo, vmo, uhtr, and vhtr diagnostics.