Remaining quantity now added properly
A bug in the original logic occurred where the remaining quantity from the
source column would be added to the last cell in the destination column if
all the source cells had been iterated through, but some small dh remained
(probably due to roundoff). To ensure that this does not happen, the
residual quantity is added to the last cell not equal to the non-missing
value.