Move regridding interpolation to separate module
This is the first commit in a series to detangle the regridding module
of MOM, and move features (such as separate regridding schemes) into
their own modules.

Here we move the interpolation-specific code, `regridding_set_ppolys`,
`interpolate_grid` and the function `interpolation_scheme` to the new
`regrid_interp` module. Unfortunately, the helper routine
`rho_interfaces_col` implements its own solver on top of the
interpolation polynomials, so we export the constants related to the
Newton-Raphson solver (`NR_ITERATIONS` and `NR_TOLERANCE`) as well.