Do btcalc on the symmetric computational domain
  Added a halo update for the ubt_Cor, which allows a number of calculations in
the setup phase of the barotropic solver to only be applied over the symmetric
memory computational domain.  This in turn will subsequently allow for narrower
halo updates.  In addition, a number of openMP directives have been cleaned up
and abbreviated, which should contribute to code clarity.  All answers are
bitwise identical.