+Refactored interpolation of h to q in CorAdCalc
  Refactored the calculation of thicknesses interpolated to corner points in
CorAdCalc to first interpolate to velocity points, to facilate the application
of open boundary conditions.  This first change is mathematically equivalent but
changes answers.  In addition, the product of G%IdxBu and G%IdyBu were replaced
with G%IareaBu in the calculation of relative vorticity, which changes answers
slightly for non-Cartesian grids.  New reference solutions were added for all
of the test cases.