Convert regridding modules away from pointer members
The initial implementation of kernelised regridding routines used
pointers within the coordinate-specific control structures. This was
required because of the possibility for a higher-level routine to
update parameters with `set_regrid_params()` (or similar) after the
initialisation of regridding. The parameter updating logic has been
pushed into the separate regridding modules. In turn, some
coordinate-specific parameters have been moved out of the main
regridding control structure.