Do HyCOM-like z* surface in adaptive coordinate
We introduce the new parameter `ADAPT_DO_MIN_DEPTH`. If true, the
`ALE_COORDINATE_CONFIG` parameter defines a target Eulerian grid that
will be used if layers are too shallow.

Perhaps this should be a function of density instead, to more closely
match the actual behaviour of the HyCOM coordinate.