Kernelise adaptive coordinate into build_ndc_column
The main computation for the adaptive coordinate was still being done
in the MOM_regridding module. It's now been broken out into a separate
module in the same fashion as the other regridding coordinates. It's a
little bit more awkward, however, as a column requires some
information from adjacent columns in order to obtain its "adaptive"
behaviour.