Wrap successive regridding calls in regrid_accelerated()
For state-based and iterative coordinates (i.e. the neutral density
adaptive coordinate), it may be the case that you want to evolve the
grid by applying the regridding calculation many times. This commit
adds a `regrid_accelerated()` routine with a similar interface to
`regridding_main()` that performs regridding `n` times.