Add option to accelerate ALE grid initialisation
State-based grids, such as the neutral density adaptive coordinate,
may be inconsistent with the initial conditions used for
initialisation. We add the `REGRID_ACCELERATE_INIT` option, which, if
true performs `REGRID_ACCELERATE_ITERATIONS` regridding/remapping
steps on the initial data to determine the initial grid used for the
first timestep of the model.