(*)Fixed state%[uv] pointer assignments
  Added explicit lower array bounds in pointer assign statements to reduced-
rank array sections for elements of the ocean surface state type, so that they
will have the correct index ranges for symmetric memory and GLOBAL_INDEXING
configurations.  This specific coding construct is a Fortran-2003 extension.
This updates corrects the answers for coupled and ice/ocean models with
symmetric memory MOM6 configurations, but no reference solutions needed to be
changed because none of the impacted configurations were being tested.