Refactor remap_all_state_vars to make dxInterface optional
The changes made in this PR make the internal logic of remap_all_state_vars
more consistent. dxInterface was an input argument which is not strictly
required. It was only needed in a specific case where the 'old' algorithm
was being used to remap velocities (which are themselves optional
 arguments). As a consquence of this change, remap_all_state_vars can now
be used when just the source and destination thicknesses are known.

Summary of changes:
-dxInterface is now declared as an optional input argument
-The order of dxInterface was was reordered so that it is included with the
other optional arguments
-A fatal error is issued if in the presence of u and v,
remap_uv_using_old_alg is also set to true, but dxInterface is not supplied