Replace null characters in attribute string in check_grid_def
Apparently, the character string for netCDF attributes can sometimes be
terminated with a null character \0. When reading in the attribute, the
Fortran intrinsic TRIM returns this null character. Thus the following
logical test would return false ('foo\0' == 'foo'). This could lead to
check_grid_def from returning an error that the read 'units' attribute did
not match the 'expected units.'