+post_integrated_diagnostics and post_surface_diagnostics
  Further refactored step_MOM to improve code readability by adding two new
subroutines, post_integrated_diagnostics and post_surface_diagnostics, that
handle several groups of diagnostics.  Also changed the arguments to
post_TS_diagnostics and changed the intent to (in) of several arguments to the
diagnostics routines in MOM.F90, and which variables are used as the temporary
arrays for calculating diagnostics.  All answers are bitwise identical.