(*)Moved set_viscous_BBL out of the dynamics code
  Moved the calls to set_viscous_BBL out of the dynamics code and up into
step_MOM, eliminating 4-way redundant calls and making the various algorithms
more obviously similar.  The answers are bitwise identical with 3 of the 4
dynamic solvers, but with MOM_dynamics_unsplit the time level of thicknesses
used to calculate the bottom boundary layer properties are replaced with the
initial thicknesses (like the other schemes), rather than the average of the
initial thicknesses and the predicted thicknesses from the first (of perhaps
several) step of the dynamics.  All existing test cases are bitwise identical,
but if SPLIT=FALSE and USE_RK2=FALSE, answers will change.