(*)Corrected the barotropic mass source compensation
  Corrected a bug when DIABATIC_FIRST=True in how the compensation for the
difference between the dynamic sea surface height used by the barotropic solver
and the height that the layers would have if net mass sources were applied
gradually instead of abruptly at the start of the thermodynamic time step. This
would change answers if both SPLIT and DIABATIC_FIRST are true and the value of
BT_MASS_SOURCE_LIMIT > 0.0, but as there are currently no test cases that match
these conditions, the answers are unchanged in all of the test cases. In cases
where the answers do change, they are unambiguously better based on weaker high
frequency gravity waves arising from coupling shocks.