(*)Convert GV%Angstrom to units of H
  GV%Angstrom was intended to be a very thin thickness in the same units (H) as
the other thicknesses, but this conversion was not being done for Boussinesq
test cases.  Several test cases (e.g., double_gyre) that had previously changed
answers when H_TO_M was not set to 1 now give the same answers when the value of
H_TO_M is varied by a factor of integer powers of 2.  Answers in test cases are
unchanged.