Neutral surface discontinuous reconstuction with no boundary extrap
After an examination, the odd behavior is due to the piecewise constant
reconstruction of T and S in the top and bottom layers.

ToDo:
-Implement PPM boundary extrapolation either by replacing the current T/S
reconstruction with the functions in ALE/PPM_functions.F90 or port over the
code so that neutral diffusion code is self-contained.
-Do something other than linear interpolation of the change in density
within a layer to find the neutral surface position.
neutral surface.