Copied new coupler_types.F90 into solo_driver
  Copied the newly reformed coupler_types code from a branch of FMS into the
solo_driver and ice_solo_driver directories, but without some of the horribly
convoluted code that should be removed from this module (hopefully with the
post-warsaw FMS commit).  This adds many new interfaces for working with
coupler_types code, and it eliminates the need for the coupler_util modules.  In
the mean-time the coupled_driver version of coupler_util has also been added to
the solo_driver directory, because that module is no longer in the
coupler_types_mod module.  All answers are bitwise identical, but many new
interface have been added.