+Added flux_init_OCMIP2_CFC
  Added a new subroutine, flux_init_OCMIP2_CFC, to set the surface values
related to CFC fluxes.  This can either be called from within the MOM6 code or
from outside because it does not require that the rest of the ocean
code has been initialized yet, or that it is even on an ocean PE.  In so doing,
it removes the need for duplicated code in ocean_model_flux_init.  All answers
are bitwise identical.