(*)Added a missing line setting an OBC index
  The I-index was not being set for the thicknesses outside of eastern open
boundary segments in the portion of write_energy that calculates layer volumes.
This has now been corrected.  This change just happens not to change answers in
the circle_OBCs test case, but this appears to be a lucky happenstance based on
the order in which the OBC segments are listed in the circle_OBC test case.