Create procedures for array and scalar inputs to EOS functions
Based on the suggestion of @Hallberg-NOAA, the array and scalar versions
of the EOS functions were wrapped into interfaces. For each of the
available equations of state, the array and scalar versions were also put
behind interfaces. For the Wright EOS, the first and second deriviative
scalar variants of the routines call the array variant by promoting
the scalar T, S, and P to 1-element arrays and demoting the 1-element
output arrays to scalars. Answers not expected to change.