(*)Added calls to open_boundary_test_extern_h
  Added debugging calls in step_MOM_dyn_split_RK2 for the open boundary
conditions that change the external values of thicknesses and pressure-force
driven accelerations if DEBUG_OBC is True.  With this initial commit, setting
DEBUG_OBC to true does change answers, but a series of commits that are all
lined up will correct this behaviour.  When DEBUG_OBC is false, all answers
are bitwise identical.