Bugfix when using passive tracers with diabatic driver but no fluxes
In TracerBoundaryFluxesInOut, a return-check for fluxes%sw was commented
out because passive tracers probably would never flux with shortwave
radiation. However, this also protected against using the routine if
freshwater fluxes were not needed. In the tracer_mixing test case when a
dye tracer was used, a runtime error was triggered because it would try to
access fluxes%netMassOut and fluxes%netMassIn which were (correctly) never
allocated. Now, the routine checks that these fields exist. If not, the
subroutine returns.