Fix for nonzero remapped vertical diffusion tendencies
As raised in https://github.com/NOAA-GFDL/MOM6-examples/issues/162,
column integrals of the vertical diffusive tendences did not sum to zero
when remapped from the model's native coordinate. Another call to
diag_update_remap_grids was added after the call to
applyBoundaryFluxesInOut (which changes the layer thicknesses due to
freshwater fluxes) and the checks for negative thickness.