+Added 'F' input_filename option
  Added the option to set the input_filename to 'F' in MOM_input_nml to use the
presence or absence of the MOM restart files to determine whether this is a new
run for the ocean.  This is a potentially risky mode of operation, but it
matches the use for other FMS component models.  These changes also required the
addition of the new public interfaces restart_files_exist, determine_is_new_run
and is_new_run in the MOM_restart module, as well as the separation of the code
to open the restart files out into a new private function, open_restart_units.
All answers are bitwise identical.