Add optional check for dt in ALE_main for diagnostic remap
ALE_main is called both during initialization and in the main loop. In the
flow_downslope/rho test case (and I suspect in others using the DOME
initialization), this leads to a segfault because the diagnostic grid has
yet to be initialized. Adding a check for the presence of dt resolves this
issue.