Fix T/S advective and horizontal diffusive fluxes and change T units
The advection and diffusive diagnostics are all extensive fields and so the
v_extensive flag needs to be set to true for the remapped version of these
fields to be done correctly. Per
https://github.com/NOAA-GFDL/MOM6-examples/issues/167, the units for
temperature have been changed to W/m^2 by including a conversion factor in
the register_diag_field call. If Boussinesq = .false., a warning is thrown
that the diagnostics are converted to Watts using a Boussinesq
approximation.