Apply conversion field for diagnostics only on non-missing values
If a diagnostic had a conversion factor associated with it, every value of
the array would be multiplied and then passed to send_data. However
because the remapping routines already fill the arrays with missing values,
this means that the places to be masked would actually have the value
missing_value*conversion leading to the output not being masked correctly.
Now, the locfield array is set to a missing value if the input field is,
and otherwise is the the value in the original array multiplied by the
conversion factor.