+Replaced calls to read_data with MOM_read_data
  Replaced the calls to the fms_io routine with calls to MOM_read_data, which
can then be overloaded to allow for an automatic rotation or other operations
based on information in the MOM_domains type.  Also added a new overloaded
version of MOM_read_data for 4-d arrays.  All answers are bitwise identical,
but a new extension of the read_data interface has been added.