+Added MOM_read_vector
  Added a new interface, MOM_read_vector, to read a pair of components of a
vector or a scalar pair.  This replaces a pair of MOM_read_data calls and will
allow for an automatic rotation or swap of the two components.  This new
interface is used in a number of locations.  All answers are bitwise identical,
but a new interface has been added to MOM_io.