Override T and S for remapped transport diagnostics
The remapping was working as expected for remapping transport diagnostics
onto a coordinate that does not depend on T/S of the model. However, for
transport on state-dependent coordinates (e.g sigma-2), the target
coordinate was being built using the pre-dynamics thicknesses, but the
current T/S.

This commit adds two optional arguments in diag_update_remap_grids to
override T and S. The override is currently only needed for the transport
diagnostics.