Revert "Change timestep used to diagnose frazil heat flux"
This reverts commit 8050108be75c734c0d2636dabb99a2ae36c38b4e.

The previous commit changed the timestep used for the heat flux from
0.5*dt_therm to dt_therm. However, this is inconsistent with the way
that the time-weighted averaging of the diagnostic manager works.

Define A1 and A2 as the change in heat for a cell which, for simplicity,
assume does not change thickness. The diagnostic manager calculates
averages as a weighted mean over the requested averaging interval (in
this case dt_therm). In the case of two averaging intervals of equal
length the weight for each term is 0.5*dt_therm/dt_therm = 0.5.
The equation for the weighted-average tendency is then
[A1/(0.5*dt_therm)]*0.5 + [A2/(0.5*dt_therm]*0.5} which simplifies
to the expected tendency (A1+A2)/dt_therm.