+Rename sfc to p_bot for non-Boussinesq restarts
  Renamed the restart variable sfc to p_bot when run in non-Boussinesq mode, to
reflect that the meaning (and likely the units) of the 2-d field eta change
between the two modes, and to avoid accidentally using one field to initialize
the other.  Because this is a non-mandatory restart variable, failure to include
this field in a restart file is automatically recovered from.  All answer are
bitwise identical in all test cases.