+THICKNESS_CONFIG="list" & convert non-Bous ICs
  Added a new option, THICKNESS_CONFIG="list", to initialize the interfaces via
a list of depths in a netcdf file, specified by INTERFACE_IC_FILE, in the
variable given by INTERFACE_IC_VAR.  Also changed the default for
CONVERT_THICKNESS_UNITS to be the opposite of Boussinesq, reflecting the
existing practice in all of the non-Boussinesq test cases, and to avoid naive
mistakes when changing from Boussinesq to non-Boussinesq options.  All answer in
the test cases are bitwise identical, but multple MOM_parameter_doc files
change.