Adds the option to balance a diffusive salt flux with a heat flux
As prescribed in Griffies et al. [1998], one can enforce the condition that
diffusion does not change locally referenced potential density if you
balance the buoyancy flux of heat or salt with the other. This commit
implements this by calculating a heat flux based on the salt flux. A
new runtime parameter NDIFF_COMP_FLUX = True enables this, but is by
default false.