Add enable averaging before ALE routines
Commit 6da5f6b inadvertently disabled posting of tendency terms due to
vertical remapping because the subroutine 'diabatic' handles the averaging
intervals for itself due to the frazil diagnostic needing to be posted
twice. All post_data calls in ALE_main were thus called while averaging
was disabled. To fix this, the call to enable_averaging is moved from the
top of step_MOM_thermo and down to where the ALE routines are called. This
resolves issue #654